requirements.txt
google-genai>=1.73.0
google-cloud-aiplatform>=1.44.0
pydantic>=2.0.0
pytest>=9.0.0
requests>=2.0.0
python-dotenv>=1.0.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.5.0

# Optional for API/mock server and visualization scripts
fastapi>=0.100.0
uvicorn>=0.30.0
networkx>=3.0.0
plotly>=6.0.0
